+++ mjx/mujoco/mjx/_src/io_test.py
@@ -434,5 +434,20 @@ class DataIOTest(parameterized.TestCase):
     self.assertEqual(d_2.contact.frame.shape, (1, 9))
     np.testing.assert_allclose(d_2.contact.frame, d.contact.frame)
 
+  def test_make_matches_put(self):
+    """Test that make_data produces a pytree that matches put_data."""
+
+    m = mujoco.MjModel.from_xml_string(_MULTIPLE_CONSTRAINTS)
+    d = mujoco.MjData(m)
+    mujoco.mj_step(m, d, 2)
+    dx = mjx.put_data(m, d)
+
+    step_fn = lambda d: d.replace(time=d.time + 1)
+    step_fn_jit = jax.jit(step_fn).lower(dx).compile()
+
+    # placing an MjData onto device should yield the same treedef mjx.Data as
+    # calling make_data.  they should be interchangeable for jax functions:
+    step_fn_jit(mjx.make_data(m))
+
 if __name__ == '__main__':
   absltest.main()